Fix OpenBSD build (#2080)
- Fix unclosed OBJ in core.cc when target is OpenBSD
- `to_real_path` changes:
  - Canonicalize result to mimic `realpath` behavior
  - Partially implement `to_real_path` for OpenBSD
  - Add more tests for `to_real_path`.
- Build system improvements:
  - Use same i18n include/link logic for all platforms
  - Replace `HAVE_SOME_SOUNDCARD_H` and `HAVE_LINUX_SOUNDCARD_H` with `HAVE_SOUNDCARD_H`
    and use other defines (__linux__) to guard include
  - Replace `<XXX>_LIBRARIES` with `<XXX>_LINK_LIBRARIES` for linking.
    - Libraries weren't in default linker search path on OpenBSD
- OpenBSD specific:
  - Fix bsdapm.cc
  - Remove diskio stats stubs
  - Stub get_battery_short_status
  - Remove `-stdlib=libc++` build flag on OpenBSD

+++ src/openbsd.h
@@ -7,16 +7,19 @@
 #include <sys/param.h>
 #include <sys/sensors.h>
 #include <sys/sysctl.h>
+#include <sys/types.h>
+#include <sys/mount.h>
+
 #include "common.h"
 
 void parse_obsd_sensor(struct text_object *, const char *);
-void print_obsd_sensors_temp(struct text_object *, char *, int);
-void print_obsd_sensors_fan(struct text_object *, char *, int);
-void print_obsd_sensors_volt(struct text_object *, char *, int);
+void print_obsd_sensors_temp(struct text_object *, char *, unsigned int);
+void print_obsd_sensors_fan(struct text_object *, char *, unsigned int);
+void print_obsd_sensors_volt(struct text_object *, char *, unsigned int);
 void get_obsd_vendor(struct text_object *, char *buf,
-                     size_t client_buffer_size);
+                     unsigned int client_buffer_size);
 void get_obsd_product(struct text_object *, char *buf,
-                      size_t client_buffer_size);
+                     unsigned int client_buffer_size);
 
 #if defined(i386) || defined(__i386__)
 typedef struct apm_power_info *apm_info_t;
